656 - Terms and conditions
2 new config elements are available.
One for enabling the T&C check and one for the actual text

The text is suppressed in the config but can be retrieved from
/api/termsAndConditions

When the T&C is enabled,  create-user requests will need to set the
`acceptedTC` parameter in the create-user request to true.
When a user signs in via an external authentication strategy the T&C
can be accepted by setting the `acceptedTC` parameter to `true` in the
update-user request.

All non-anonymous POSTS requests not going to /api/auth or /api/user
will be checked.